requirements.txt
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
PyYAML==6.0.2
scipy==1.11.4
python-dotenv==1.1.0
pymssql==2.3.4
openpyxl==3.1.5
tqdm==4.67.1
cassandra-driver==3.29.2
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
opentelemetry-api>=1.14.0
opentelemetry-sdk>=1.14.0
opentelemetry-instrumentation-fastapi>=0.55b1
opentelemetry-instrumentation-requests>=0.55b1
opentelemetry-exporter-otlp-proto-grpc>=1.14.0
opentelemetry-exporter-console>=1.14.0


